+++ packages/gov-ui-kit/src/modules/components/wallet/wallet.stories.tsx
@@ -25,24 +25,55 @@ type Story = StoryObj<typeof Wallet>;
 export const Default: Story = {};
 
 /**
- * Example usage of a connected Wallet component with only address.
+ * Example usage of the Wallet component with only address.
  */
-export const WithAddress: Story = {
+export const Address: Story = {
+    args: {
+        user: { address: '0x1234567890123456789012345678901234567890' },
+    },
+};
+
+/**
+ * The Wallet component resolves and displays the linked ENS name and ENS avatar of the specified address.
+ */
+export const ResolvedEnsData: Story = {
+    args: {
+        user: { address: '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045' },
+    },
+};
+
+/**
+ * Example usage of the Wallet component with a custom name.
+ */
+export const CustomName: Story = {
     args: {
         user: {
-            address: '0x1234567890123456789012345678901234567890',
+            address: '0xdafbd7d63cee88d73a51592b42f27f7fd6ab7722',
+            name: 'Aragon DAO',
         },
     },
 };
 
 /**
- * Example usage of a connected Wallet component with additional ENS handle provided.
+ * Example usage of the Wallet component with a custom avatar.
  */
-export const WithEnsName: Story = {
+export const CustomAvatar: Story = {
     args: {
         user: {
-            address: '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045',
-            name: 'vitalik.eth',
+            address: '0xdafbd7d63cee88d73a51592b42f27f7fd6ab7722',
+            avatarSrc: 'https://aragon-1.mypinata.cloud/ipfs/Qmbtf1tLJrNFDgxyqXxaYVZt9pifarukuhNve7aMXLP9nh',
+        },
+    },
+};
+
+/**
+ * Example usage of the Wallet component with a long name.
+ */
+export const LongName: Story = {
+    args: {
+        user: {
+            address: '0xdafbd7d63cee88d73a51592b42f27f7fd6ab7722',
+            name: 'alongensname.eth',
         },
     },
 };