+++ packages/gov-ui-kit/src/modules/components/wallet/wallet.test.tsx
@@ -1,5 +1,6 @@
 import { type QueryClient } from '@tanstack/react-query';
 import { render, screen } from '@testing-library/react';
+import { sepolia } from 'viem/chains';
 import * as wagmi from 'wagmi';
 import { OdsModulesProvider } from '../odsModulesProvider';
 import { Wallet, type IWalletProps } from './wallet';
@@ -10,8 +11,8 @@ jest.mock('../member', () => ({
 
 jest.mock('../../utils/addressUtils', () => ({
     addressUtils: {
-        getChecksum: jest.fn((address) => address),
-        truncateAddress: jest.fn((address) => `${address?.slice(0, 4)}…${address?.slice(-4)}`),
+        getChecksum: (address: string) => address,
+        truncateAddress: (address: string) => `${address?.slice(0, 4)}…${address?.slice(-4)}`,
     },
 }));
 
@@ -23,10 +24,10 @@ describe('<Wallet /> component', () => {
             ...props,
         };
 
-        return render(
+        return (
             <OdsModulesProvider queryClient={queryClient}>
                 <Wallet {...completeProps} />
-            </OdsModulesProvider>,
+            </OdsModulesProvider>
         );
     };
 
@@ -39,58 +40,56 @@ describe('<Wallet /> component', () => {
     });
 
     it('renders connect button when disconnected', () => {
-        createTestComponent();
+        render(createTestComponent());
         const button = screen.getByRole('button');
         expect(button).toHaveTextContent('Connect');
     });
 
+    it('renders a loading indicator when loading the user ENS name', () => {
+        const user = { address: '0x0987654321098765432109876543210987654321' };
+        useEnsNameMock.mockReturnValue({ isLoading: true } as wagmi.UseEnsNameReturnType);
+        render(createTestComponent({ user }));
+        expect(screen.getByTestId('stateSkeletonBar')).toBeInTheDocument();
+    });
+
     it('renders member avatar when connected', () => {
-        const user = {
-            address: '0x0987654321098765432109876543210987654321',
-        };
-        createTestComponent({ user });
+        const user = { address: '0x0987654321098765432109876543210987654321' };
+        render(createTestComponent({ user }));
         expect(screen.getByTestId('member-avatar-mock')).toBeInTheDocument();
     });
 
-    it('renders user address when connected and user has no ENS name linked', () => {
-        const user = {
-            address: '0x0987654321098765432109876543210987654321',
-        };
-        useEnsNameMock.mockReturnValue({
-            data: null,
-            isLoading: false,
-        } as wagmi.UseEnsNameReturnType);
+    it('renders truncated user address when connected and user has no ENS name linked', () => {
+        const user = { address: '0x0987654321098765432109876543210987654321' };
+        useEnsNameMock.mockReturnValue({ data: null, isLoading: false } as wagmi.UseEnsNameReturnType);
 
-        createTestComponent({ user });
+        render(createTestComponent({ user }));
         expect(screen.getByText('0x09…4321')).toBeInTheDocument();
     });
 
-    it('renders linked ENS name when connected and no user name provided', () => {
-        const user = {
-            address: '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045',
-        };
-        useEnsNameMock.mockReturnValue({
-            data: 'vitalik.eth',
-            isLoading: false,
-        } as wagmi.UseEnsNameReturnType);
+    it('resolves and renders the linked ENS name when connected and no user name provided', () => {
+        const user = { address: '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045' };
+        useEnsNameMock.mockReturnValue({ data: 'vitalik.eth', isLoading: false } as wagmi.UseEnsNameReturnType);
 
-        createTestComponent({ user });
+        render(createTestComponent({ user }));
+        expect(useEnsNameMock).toHaveBeenCalledWith(expect.objectContaining({ query: { enabled: true } }));
         expect(screen.getByText('vitalik.eth')).toBeInTheDocument();
         expect(screen.queryByText('0xd8…6045')).not.toBeInTheDocument();
     });
 
-    it('renders user name provided when connected and does not fetch linked ENS name', () => {
-        const user = {
-            address: '0x0987654321098765432109876543210987654321',
-            name: 'vitalik.eth',
-        };
-        useEnsNameMock.mockReturnValue({
-            data: 'vitalikeviltwin.eth',
-            isLoading: false,
-        } as wagmi.UseEnsNameReturnType);
+    it('renders user name provided when connected and does not resolve the ENS name', () => {
+        const user = { address: '0x0987654321098765432109876543210987654321', name: 'vitalik.eth' };
+        useEnsNameMock.mockReturnValue({ data: 'vitalikeviltwin.eth', isLoading: false } as wagmi.UseEnsNameReturnType);
 
-        createTestComponent({ user });
+        render(createTestComponent({ user }));
+        expect(useEnsNameMock).toHaveBeenCalledWith(expect.objectContaining({ query: { enabled: false } }));
         expect(screen.getByText('vitalik.eth')).toBeInTheDocument();
         expect(screen.queryByText('0x09…4321')).not.toBeInTheDocument();
+    });
+
+    it('supports custom chainId and wagmi configurations', () => {
+        const chainId = 137;
+        const wagmiConfig = { chains: [sepolia] } as unknown as wagmi.Config;
+        render(createTestComponent({ chainId, wagmiConfig }));
+        expect(useEnsNameMock).toHaveBeenCalledWith(expect.objectContaining({ chainId, config: wagmiConfig }));
     });
 });