+++ packages/gov-ui-kit/src/modules/components/wallet/wallet.tsx
@@ -1,12 +1,12 @@
 import classNames from 'classnames';
 import { useEnsName } from 'wagmi';
 import { StateSkeletonBar } from '../../../core';
-import { type ICompositeAddress } from '../../types';
+import { type ICompositeAddress, type IWeb3ComponentProps } from '../../types';
 import { addressUtils } from '../../utils';
 import { MemberAvatar } from '../member';
 import { useOdsModulesContext } from '../odsModulesProvider';
 
-export interface IWalletProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+export interface IWalletProps extends React.ButtonHTMLAttributes<HTMLButtonElement>, IWeb3ComponentProps {
     /**
      * The connected user details.
      */
@@ -14,42 +14,37 @@ export interface IWalletProps extends React.ButtonHTMLAttributes<HTMLButtonEleme
 }
 
 export const Wallet: React.FC<IWalletProps> = (props) => {
-    const { user, className, ...otherProps } = props;
-    const { data: ensName, isLoading: ensLoading } = useEnsName({
-        address: user != null ? addressUtils.getChecksum(user.address) : undefined,
-        query: { enabled: user?.name != null },
-    });
-    const resolvedUserHandle = user?.name ?? ensName ?? addressUtils.truncateAddress(user?.address);
+    const { user, className, chainId, wagmiConfig, ...otherProps } = props;
 
     const { copy } = useOdsModulesContext();
 
+    const { data: ensName, isLoading: isEnsLoading } = useEnsName({
+        address: user != null ? addressUtils.getChecksum(user.address) : undefined,
+        query: { enabled: user != null && user.name == null },
+        chainId,
+        config: wagmiConfig,
+    });
+    const resolvedUserHandle = user?.name ?? ensName ?? addressUtils.truncateAddress(user?.address);
+
     const buttonClassName = classNames(
-        'flex items-center rounded-full border border-neutral-100 bg-neutral-0 text-neutral-500 transition-all',
-        'hover:border-neutral-200',
+        'flex items-center gap-3 rounded-full border border-neutral-100 bg-neutral-0 text-neutral-500 transition-all',
+        'hover:border-neutral-200 active:bg-neutral-50 active:text-neutral-800',
         'focus:outline-none focus-visible:ring focus-visible:ring-primary focus-visible:ring-offset',
-        'active:bg-neutral-50 active:text-neutral-800',
-        {
-            'px-4 py-2.5': user == null,
-            'p-1 md:pl-4': user != null,
-        },
+        { 'px-4 py-2.5': user == null },
+        { 'p-1 md:pl-4': user != null },
         className,
     );
 
     return (
         <button className={buttonClassName} {...otherProps}>
             {!user && copy.wallet.connect}
-
-            {user && (
-                <>
-                    <div
-                        title={user.name ?? user.address}
-                        className="hidden min-w-0 max-w-24 items-center truncate md:mr-3 md:flex"
-                    >
-                        {ensLoading ? <StateSkeletonBar size="lg" /> : resolvedUserHandle}
-                    </div>
-                    <MemberAvatar size="lg" ensName={user.name} address={user.address} />
-                </>
+            {user && isEnsLoading && <StateSkeletonBar size="lg" width={56} />}
+            {user && !isEnsLoading && (
+                <span title={user.name ?? ensName ?? user.address} className="hidden max-w-24 truncate md:block">
+                    {resolvedUserHandle}
+                </span>
             )}
+            {user && <MemberAvatar size="lg" ensName={user.name} address={user.address} avatarSrc={user.avatarSrc} />}
         </button>
     );
 };